+++ bookmarks/views/contexts.py
@@ -6,6 +6,7 @@ from django.core.paginator import Paginator
 from django.db import models
 from django.http import Http404
 from django.urls import reverse
+from django.utils.translation import gettext_lazy as _
 
 from bookmarks import queries, utils
 from bookmarks.forms import BookmarkSearchForm
@@ -298,7 +299,7 @@ class BookmarkListContext:
 
 
 class ActiveBookmarkListContext(BookmarkListContext):
-    list_title = "Bookmarks"
+    list_title = _("Bookmarks")
     search_mode = ""
     bulk_edit_enabled = True
     bulk_edit_disabled_actions = "bulk_unarchive"
@@ -306,7 +307,7 @@ class ActiveBookmarkListContext(BookmarkListContext):
 
 
 class ArchivedBookmarkListContext(BookmarkListContext):
-    list_title = "Archived bookmarks"
+    list_title = _("Archived bookmarks")
     search_mode = "archived"
     bulk_edit_enabled = True
     bulk_edit_disabled_actions = "bulk_archive"
@@ -314,7 +315,7 @@ class ArchivedBookmarkListContext(BookmarkListContext):
 
 
 class SensitiveArchivedBookmarkListContext(BookmarkListContext):
-    list_title = "Sensitive archived bookmarks"
+    list_title = _("Sensitive archived bookmarks")
     search_mode = "archived-sensitive"
     bulk_edit_enabled = True
     bulk_edit_disabled_actions = "bulk_archive"
@@ -322,7 +323,7 @@ class SensitiveArchivedBookmarkListContext(BookmarkListContext):
 
 
 class SensitiveBookmarkListContext(BookmarkListContext):
-    list_title = "Sensitive bookmarks"
+    list_title = _("Sensitive bookmarks")
     search_mode = "sensitive"
     bulk_edit_enabled = True
     bulk_edit_disabled_actions = "bulk_unarchive"
@@ -330,7 +331,7 @@ class SensitiveBookmarkListContext(BookmarkListContext):
 
 
 class SharedBookmarkListContext(BookmarkListContext):
-    list_title = "Shared bookmarks"
+    list_title = _("Shared bookmarks")
     search_mode = "shared"
     bulk_edit_enabled = False
     bulk_edit_disabled_actions = ""