+++ bookmarks/views/settings.py
@@ -12,6 +12,7 @@ from django.http import HttpResponse, HttpResponseRedirect
 from django.shortcuts import render
 from django.urls import reverse
 from django.utils import timezone
+from django.utils.translation import gettext as _
 
 from bookmarks.forms import GlobalSettingsForm, UserProfileForm
 from bookmarks.models import (
@@ -73,13 +74,13 @@ def update(request: HttpRequest):
         if "update_global_settings" in request.POST:
             update_global_settings(request)
             messages.success(
-                request, "Global settings updated", "settings_success_message"
+                request, _("Global settings updated"), "settings_success_message"
             )
         if "refresh_favicons" in request.POST:
             tasks.schedule_refresh_favicons(request.user)
             messages.success(
                 request,
-                "Scheduled favicon update. This may take a while...",
+                _("Scheduled favicon update. This may take a while..."),
                 "settings_success_message",
             )
         if "create_missing_html_snapshots" in request.POST:
@@ -87,12 +88,13 @@ def update(request: HttpRequest):
             if count > 0:
                 messages.success(
                     request,
-                    f"Queued {count} missing snapshots. This may take a while...",
+                    _("Queued %(count)s missing snapshots. This may take a while...")
+                    % {"count": count},
                     "settings_success_message",
                 )
             else:
                 messages.success(
-                    request, "No missing snapshots found.", "settings_success_message"
+                    request, _("No missing snapshots found."), "settings_success_message"
                 )
 
     return HttpResponseRedirect(reverse("linkding:settings.general"))
@@ -106,7 +108,7 @@ def update_profile(request: HttpRequest):
     form = UserProfileForm(request.POST, instance=profile)
     if form.is_valid():
         form.save()
-        messages.success(request, "Profile updated", "settings_success_message")
+        messages.success(request, _("Profile updated"), "settings_success_message")
         # Load missing favicons if the feature was just enabled
         if profile.enable_favicons and not favicons_were_enabled:
             tasks.schedule_bookmarks_without_favicons(request.user)
@@ -118,7 +120,7 @@ def update_profile(request: HttpRequest):
 
     messages.error(
         request,
-        "Profile update failed, check the form below for errors",
+        _("Profile update failed, check the form below for errors"),
         "settings_error_message",
     )
     return general(request, 422, {"form": form})